1) First implementation of the network learning how to read handwritten numerals. The code currently trains by the 60,000 MNIST images and labels (which are included here). 2) The MNIST images are broken up into batches of 64 at a time, accumulating gradients, and then updating weights and biases after each batch. 3) The NetworkRunner class now has a switch statement in the main() function. You can specify up top which activity to train (XOR, 4 functions, or number reading) and then the network will learn accordingly. 4) Added a new Layer type - SoftmaxLayer - implemented only at the output Layer. This is useful for the MNIST dataset, as earlier runs for me with SigmoidLayer resulted in 8 never quite getting recognized. At this point, thicker numbers without holes tend not to get recognized with the existing Network setup, but I'll play around with Layer sizes to experiment. SoftmaxLayer is used here in conjunction with categorical cross-entropy (as the two together end up with simple (actual-expected) calculation for figuring out the delta during back propagation.
Prior to that:
JFreeCharting and 4 new functions
Trained the network with 4 new functions: sin(x), cos(x), x^2, and
abs(x).  All at the same time!  The training data is dynamically
generated and fed into the system each epoch.  Then after training is
complete, an additional set of random inputs is generated and used to
produce and verify outputs based on the expected output for each input.
For these functions, the output layer was switched to Tanh (since we
need to allow negative values in the output).

Also implemented JFreeCharting to show the XY Series Collection of data,
which is the MSE (mean squared error) calculation for each output node
for each epoch.  The 4 outputs are charted and labeled, so you can see
how quickly the network learns and reduces errors.

Prior to that:
File I/O and New Training
Added file IO to both save and retrieve Network setups, using Gson.
Every training run will save the final weights and biases in a JSON
file, which can be used to pre-load a Network (of the same architecture
for now) with those weights and biases to run additional inputs through.

Also modified the training from just XOR to training 4 logic gates at
once: AND, OR, NAND, and XOR.  All for 2 bits for now.