Most changes implemented here are related to trying to get this network to learn what a turtle is.   Since the last commit, here's what I've tried:
Originally I tried with a Sigmoid/BCE output Layer. When that didn't
work, I added validation steps in the process to track F1 data
(thresholds at first, argmax later). Then I increased the image set -
from 50 turtles/50 others to 215 of each. Then I increased augmentation
(to about 6,000 images total). Then I modified the image set again to
have only turtles of similar styles (clipart basically) and the "other"
images being mostly turtle-like critters (like ladybugs, clams,
frisbees, etc).  Then I augmented the new images (to get to about 12,000
total). That didn't work, so I revamped Layer/Loss to include
Softmax/CCE (which was a major overhaul because of Vector vs. Scalar).
That didn't work either. So far, with all my attempts, the network
"learns" in the sense that loss decreases and is smooth every single
time (implying the network architecture and wiring is all good).  But it
doesn't categorize well, as the F1 score has only reached 0.6 a few
times.  So the next step is to change the training images again - keep
the turtles as-is, but pick a single type of "other" and get a similar
number of images of that, then train the network on it.  The images
aren't your concern here.  But the code is!  So here are the latest
changes to the code since the last commit:
* Extracted Losses to be independent of any output Layers.  Now when you
define the output Layer activation type, you also need to inject the
Loss function into it.  Right now there is CCE (vector based), BCE, and
MSE (scalar based) Loss functions.
* Added Softmax properly (previously it was neuron-based instead of
layer-based).  Right now the only Loss that can work with Softmax is
CCE.
* Added a Threshold and ArgMax analyzer, which should be utilized when
doing a validation pass - they print helpful info about how training is
going, such as F1, recall, precision, and accuarcy.
* Added all the setup for turtle image reading, as that is the next goal
of this network.  You need to provide your own turtles, of course.